pkg/bb: do not return an error if no main can be found
build constraints can result in there being no main, e.g.,
GOOS=plan9 u-root cmds/core/d*
should not be an error just because plan 9 does not have or want dmesg.

Why do this? Because linux-ism's keep creeping back into our code. The easiest way to
avoid them is to build everything with GOOS=plan9, and let build constraints elide
the ones that are not supported.

For that to work, u-root has to ignore packages that have no main after build
constraints are applied, not error out on them.

This is a bit of a hack until we can get the new gobusybox code in use.

+++ pkg/bb/bb.go
@@ -146,8 +146,12 @@ func BuildBusybox(env golang.Environ, pkgs []string, noStrip bool, binaryPath st
 		seenPackages[basePkg] = true
 
 		// TODO: use bbDir to derive import path below or vice versa.
-		if err := RewritePackage(env, pkg, "github.com/u-root/u-root/pkg/bb/bbmain", importer); err != nil {
+		ok, err := RewritePackage(env, pkg, "github.com/u-root/u-root/pkg/bb/bbmain", importer)
+		if err != nil {
 			return err
+		}
+		if !ok {
+			continue
 		}
 
 		bbPackages = append(bbPackages, path.Join(pkg, ".bb"))
@@ -272,7 +276,7 @@ func ParseAST(files []string) (*token.FileSet, *ast.Package, error) {
 
 	// Did we parse anything?
 	if len(p.Files) == 0 {
-		return nil, nil, fmt.Errorf("no valid `main` package files found in %v", files)
+		return nil, nil, nil
 	}
 	return fset, p, nil
 }
@@ -289,23 +293,26 @@ func SrcFiles(p *build.Package) []string {
 // RewritePackage rewrites pkgPath to be bb-mode compatible, where destDir is
 // the file system destination of the written files and bbImportPath is the Go
 // import path of the bb package to register with.
-func RewritePackage(env golang.Environ, pkgPath, bbImportPath string, importer types.Importer) error {
+func RewritePackage(env golang.Environ, pkgPath, bbImportPath string, importer types.Importer) (bool, error) {
 	buildp, err := env.Package(pkgPath)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	p, err := NewPackage(filepath.Base(buildp.Dir), buildp.ImportPath, SrcFiles(buildp), importer)
 	if err != nil {
-		return err
+		return false, err
+	}
+	if p == nil {
+		return false, nil
 	}
 	dest := filepath.Join(buildp.Dir, ".bb")
 	// If .bb directory already exists, delete it. This will prevent stale
 	// files from being included in the build.
 	if err := os.RemoveAll(dest); err != nil {
-		return fmt.Errorf("error removing stale directory %q: %v", dest, err)
+		return false, fmt.Errorf("error removing stale directory %q: %v", dest, err)
 	}
-	return p.Rewrite(dest, bbImportPath)
+	return true, p.Rewrite(dest, bbImportPath)
 }
 
 // NewPackage gathers AST, type, and token information about a command.
@@ -315,6 +322,10 @@ func NewPackage(name string, pkgPath string, srcFiles []string, importer types.I
 	fset, astp, err := ParseAST(srcFiles)
 	if err != nil {
 		return nil, err
+	}
+
+	if fset == nil {
+		return nil, nil
 	}
 
 	p := &Package{
